feat: complete

+++ server/src/controllers/like.controller.ts
@@ -1,3 +1,4 @@
+import { badRequest } from "@hapi/boom"
 import { Request, Response } from "express"
 import { Types } from "mongoose"
 import { z } from "zod"
@@ -16,7 +17,7 @@ export const getPostLikes = async (req: Request, res: Response) => {
   await checkPostExists(postId)
   const likes = await Like.find({
     postId: postId
-  }).lean()
+  }).populate('userId', 'name').lean()
   return res.status(200).json(likes)
 }
 
@@ -29,11 +30,18 @@ export const likePost = async (req: Request, res: Response) => {
   const { params } = await zParse(likePostSchema, req)
   const { postId } = params
   await checkPostExists(postId)
-  const like = await Like.create({
+  const like = await Like.exists({
     userId: req.user?._id,
     postId: postId
   })
-  return res.status(201).json(like)
+  if (like) {
+    return badRequest("Post already liked")
+  }
+  const newLike = await Like.create({
+    userId: req.user?._id,
+    postId: postId
+  })
+  return res.status(201).json(await newLike.populate('userId', 'name'))
 }
 
 export const unlikePost = async (req: Request, res: Response) => {
@@ -45,9 +53,16 @@ export const unlikePost = async (req: Request, res: Response) => {
   const { params } = await zParse(unlikePostSchema, req)
   const { postId } = params
   await checkPostExists(postId)
-  const like = await Like.findOneAndDelete({
+  const like = await Like.exists({
     userId: req.user?._id,
     postId: postId
   })
-  return res.status(200).json(like)
+  if (!like) {
+    return badRequest("Post not liked")
+  }
+  await Like.findOneAndDelete({
+    userId: req.user?._id,
+    postId: postId
+  })
+  return res.status(200).json({ message: "Post like removed" })
 }
